only support sl-tinkergraph
