--- cpgloader-janusgraph/src/main/scala/io/shiftleft/cpgloading/janusgraph/CpgLoader.scala
@@ -1,11 +1,0 @@
-package io.shiftleft.cpgloading.janusgraph
-
-import io.shiftleft.cpgloading.CpgLoaderBase
-
-/* singleton instance for convenience */
-object CpgLoader extends CpgLoader
-
-/** Load cpg proto (typically cpg.bin.zip) into a new Janusgraph instance */
-class CpgLoader extends CpgLoaderBase {
-  override protected def builder = new ProtoToCpg
-}
